refactor: Record(Room migration, Kotlin) - SQLiteOpenHelper에서 Room을 사용하는 구조로 변경, 데이터 Migration 작업 수행 - MVVM 구조로 변경 - ViewPager를 ViewPager2로 변경 - 기존 탭 3개에서 2개로 변경(일정별 기록 탭 제거 - 캘린더로 충분)

+++ src/main/java/com/health/myapplication/listener/DataListener.java
@@ -1,5 +1,5 @@
 package com.health.myapplication.listener;
 
 public interface DataListener {
-    public void onPositiveClicked(String time, String name, int set, int rep, float weight); //for training data
+    public void onPositiveClicked(String time, String name, int set, int rep, double weight); //for training data
 }
